# introduces a faster query to tuple GC