we've noticed that even after https://github.com/authzed/spicedb/pull/1550 was introduced and made the nested GC SELECT query faster, deletions are still causing spikes in postgres
and Aurora instances.

The proposed query seemed 1 order of magnitude less cost that the current query on a test PG 15.6 with 5M dead tuples and 5M alive tuples.

## Before

```sql
WITH rows AS (SELECT namespace, object_id, relation, userset_namespace, userset_object_id, userset_relation, created_xid, deleted_xid FROM relation_tuple WHERE deleted_xid < '901' LIMIT 1000)
DELETE FROM relation_tuple
WHERE (namespace, object_id, relation, userset_namespace, userset_object_id, userset_relation, created_xid, deleted_xid) IN (SELECT namespace, object_id, relation, userset_namespace, userset_object_id, userset_relation, created_xid, deleted_xid FROM rows);

```

```
| -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
| Delete on relation_tuple  (cost=65.94..8656.38 rows=0 width=0) (actual time=14.375..14.376 rows=0 loops=1)                                                                                                                                                                                                                                                                                                     |
|   ->  Nested Loop  (cost=65.94..8656.38 rows=1000 width=110) (actual time=1.100..14.026 rows=1000 loops=1)                                                                                                                                                                                                                                                                                                     |
|         ->  HashAggregate  (cost=65.38..75.38 rows=1000 width=184) (actual time=1.061..1.319 rows=1000 loops=1)                                                                                                                                                                                                                                                                                                |
|               Group Key: (rows.namespace)::text, (rows.object_id)::text, (rows.relation)::text, (rows.userset_namespace)::text, (rows.userset_object_id)::text, (rows.userset_relation)::text, rows.created_xid, rows.deleted_xid                                                                                                                                                                              |
|               Batches: 1  Memory Usage: 841kB                                                                                                                                                                                                                                                                                                                                                                  |
|               ->  Subquery Scan on rows  (cost=0.43..45.38 rows=1000 width=184) (actual time=0.150..0.647 rows=1000 loops=1)                                                                                                                                                                                                                                                                                   |
|                     ->  Limit  (cost=0.43..35.38 rows=1000 width=80) (actual time=0.143..0.488 rows=1000 loops=1)                                                                                                                                                                                                                                                                                              |
|                           ->  Index Scan using ix_gc_index on relation_tuple relation_tuple_1  (cost=0.43..178606.61 rows=5110010 width=80) (actual time=0.142..0.444 rows=1000 loops=1)                                                                                                                                                                                                                       |
|                                 Index Cond: (deleted_xid < '901'::xid8)                                                                                                                                                                                                                                                                                                                                        |
|         ->  Index Scan using uq_relation_tuple_living_xid on relation_tuple  (cost=0.56..8.57 rows=1 width=86) (actual time=0.012..0.012 rows=1 loops=1000)                                                                                                                                                                                                                                                    |
|               Index Cond: (((namespace)::text = (rows.namespace)::text) AND ((object_id)::text = (rows.object_id)::text) AND ((relation)::text = (rows.relation)::text) AND ((userset_namespace)::text = (rows.userset_namespace)::text) AND ((userset_object_id)::text = (rows.userset_object_id)::text) AND ((userset_relation)::text = (rows.userset_relation)::text) AND (deleted_xid = rows.deleted_xid)) |
|               Filter: (rows.created_xid = created_xid)                                                                                                                                                                                                                                                                                                                                                         |
| Planning Time: 0.674 ms                                                                                                                                                                                                                                                                                                                                                                                        |
| Execution Time: 14.464 ms                                                                                                                                                                                                                                                                                                                                                                                      |
```

## After

```sql
WITH rows AS (SELECT ctid FROM relation_tuple WHERE deleted_xid < '901' LIMIT 1000)
DELETE FROM relation_tuple
WHERE ctid IN (SELECT ctid FROM rows);
```

```
| --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
| Delete on relation_tuple  (cost=47.89..854.38 rows=0 width=0) (actual time=3.598..3.598 rows=0 loops=1)                                                                                 |
|   ->  Nested Loop  (cost=47.89..854.38 rows=1000 width=36) (actual time=2.682..3.237 rows=1000 loops=1)                                                                                 |
|         ->  HashAggregate  (cost=47.88..49.88 rows=200 width=36) (actual time=2.679..2.758 rows=1000 loops=1)                                                                           |
|               Group Key: rows.ctid                                                                                                                                                      |
|               Batches: 1  Memory Usage: 209kB                                                                                                                                           |
|               ->  Subquery Scan on rows  (cost=0.43..45.38 rows=1000 width=36) (actual time=0.193..2.503 rows=1000 loops=1)                                                             |
|                     ->  Limit  (cost=0.43..35.38 rows=1000 width=6) (actual time=0.183..2.406 rows=1000 loops=1)                                                                        |
|                           ->  Index Scan using ix_gc_index on relation_tuple relation_tuple_1  (cost=0.43..178606.61 rows=5110010 width=6) (actual time=0.182..2.363 rows=1000 loops=1) |
|                                 Index Cond: (deleted_xid < '901'::xid8)                                                                                                                 |
|         ->  Tid Scan on relation_tuple  (cost=0.00..4.01 rows=1 width=6) (actual time=0.000..0.000 rows=1 loops=1000)                                                                   |
|               TID Cond: (ctid = rows.ctid)                                                                                                                                              |
| Planning Time: 3.335 ms                                                                                                                                                                 |
| Execution Time: 3.706 ms                                                                                                                                                                |
```