@vroldanbet 2024-04-09
I was investigating if `ctid` is stable within a CTE, which led me to this post that describes [a gotcha on partitioned tables](https://dba.stackexchange.com/a/322964):

>There actually is a serious pitfall with using ctid for identifying rows to delete: in partitioned tables, ctid is not unique. It is only unique in a given partition so, if one uses it as a criterion for deletion, this will likely result in deleting the desired row as well as the otherwise totally unrelated rows from other partitions (ask me how I know :-/ ).
 
I ran the query with `tableoid` and it seems to work just fine, so we should use it.
```
Delete on relation_tuple  (cost=72.12..879.11 rows=0 width=0) (actual time=1.427..1.428 rows=0 loops=1)
"  ->  Nested Loop  (cost=72.12..879.11 rows=500 width=40) (actual time=0.569..1.157 rows=1000 loops=1)"
"        ->  HashAggregate  (cost=72.11..74.11 rows=200 width=44) (actual time=0.565..0.648 rows=1000 loops=1)"
"              Group Key: rows.tableoid, rows.ctid"
"              Batches: 1  Memory Usage: 209kB"
"              ->  Subquery Scan on rows  (cost=0.00..67.11 rows=1000 width=44) (actual time=0.024..0.407 rows=1000 loops=1)"
"                    ->  Limit  (cost=0.00..57.11 rows=1000 width=10) (actual time=0.020..0.315 rows=1000 loops=1)"
"                          ->  Seq Scan on relation_tuple relation_tuple_1  (cost=0.00..229045.70 rows=4010286 width=10) (actual time=0.020..0.272 rows=1000 loops=1)"
"                                Filter: (deleted_xid < '901'::xid8)"
"        ->  Tid Scan on relation_tuple  (cost=0.00..4.02 rows=1 width=10) (actual time=0.000..0.000 rows=1 loops=1000)"
"              TID Cond: (ctid = rows.ctid)"
"              Filter: (rows.tableoid = tableoid)"
Planning Time: 0.106 ms
Execution Time: 1.470 ms
```