@vroldanbet 2024-04-10
>@vroldanbet Is there a test we can add for GC along a partitioned table to ensure the use of the table ID fixes this potential issue?

We cannot control `tableoid` and `ctid`, those are internal columns. From the docs:

[What the docs say about `tableoid`](https://www.postgresql.org/docs/current/ddl-system-columns.html#DDL-SYSTEM-COLUMNS-TABLEOID):

>The OID of the table containing this row. This column is particularly handy for queries that select from partitioned tables or inheritance hierarchies, since without it, it's difficult to tell which individual table a row came from. The tableoid can be joined against the oid column of pg_class to obtain the table name.

[What the docs say about `ctid`](https://www.postgresql.org/docs/current/ddl-system-columns.html#DDL-SYSTEM-COLUMNS-CTID):

>The physical location of the row version within its table. Note that although the ctid can be used to locate the row version very quickly, a row's ctid will change if it is updated or moved by VACUUM FULL. Therefore ctid is useless as a long-term row identifier. A primary key should be used to identify logical rows.

Testing a scenario where two different relationships to be GCed have the same `ctid` would require us to:
- add a new type of test suite that partitions the main tables defined in the datastore. Something we should do at some point, but should be probably tackled into its own PR
- generate a state of the database where 2 CTIDs are the same across partitions. I'm not sure how _likely_ that is to happen without understanding how it's implemented. Docs say its "the location in the table", so it does not sound like "randomly generated". We'd have to create a test where we have a revision marked as deleted in one partition and a revision that is alive in another partition, both with the same CTID.

My interpretation of the document is that `tableoid, ctid` provides a globally unique internal identification to the row.